Add files via upload
This version of the Grove LSM303d 6dof sensor corrects a problem of the original Grove library for Arduino. The issue was the use of the unspecific type 'int' instead 'int16_t' in the integer transactions with the sensor. The sensed data is provided as 16 bit signed integers.
While it worked with microcontrollers as found in Uno and Mega modules, it failed with Arm and other 32bit microcontrollers, since in these 'int' would mean a 32bit integer.
C++ has provisions for specific and portable data types as int16_t, which is the solution proposed here.
A Sony Spresense module connected to the Grove LSM303d previously failed systematically and now works flawlessly.